+++ packages/orm/src/betterAuth/index.ts
@@ -1,0 +1,2 @@
+export * from './authEntities.js';
+export * from './uqlAdapter.js';